Project/requirements.txt
click==8.0.3
matplotlib==3.4.3
numpy==1.20.3
python-dotenv==0.19.2
setuptools==58.0.4
torch==1.10.1
torchvision==0.11.2
